feat: retrieve player score from database before updates

+++ server/src/player/Player.hpp
@@ -29,7 +29,7 @@ namespace game {
 
         [[nodiscard]] const std::string &getId() const { return this->_id; }
 
-        uint16_t getScore() { return this->_score; }
+        int getScore() { return this->_score; }
 
         void storeInfo(const network::ClientTCPReceivedInfo &info)
         {
@@ -66,14 +66,14 @@ namespace game {
 
         std::string getUserPass() { return this->_userPass; };
 
-        void setScore(uint16_t score) { this->_score = score; }
+        void setScore(int score) { this->_score = score; }
 
        private:
         network::UDPNetwork &_network;
         component::PlayerInput _lastInput;
         std::string _id;
         unsigned _entity = BASE_ENTITY;
-        uint16_t _score = 0;
+        int _score = 0;
         uint16_t _udpPort = 0;
         std::string _userPass;
         uint8_t _room = 0;
